0611Yib

+++ Frontend/src/routes/activity/to_do_list.jsx
@@ -97,6 +97,11 @@ export default function To_do_list() {
     // Delete order ------------------------------------------------------------------
 
     const handleDeleteClick = async (item) => {
+        const isConfirmed = window.confirm("確定要刪除此待辦事項嗎？");
+
+        if (!isConfirmed) {
+            return;
+        }
         try {
             await deleteEventsToDo(item.id);
             alert('已成功刪除待辦事項');
