+++ Frontend/src/routes/purchase/river_order.jsx
@@ -96,6 +96,11 @@ export default function River_order() {
     // Delete order ------------------------------------------------------------------
 
     const handleDeleteClick = async (item) => {
+        const isConfirmed = window.confirm("確定要刪除此待辦事項嗎？");
+
+        if (!isConfirmed) {
+            return;
+        }
         try {
             await deleteRiverOrder(item.id);
             alert('已成功刪除待辦事項');
